+++ PloidPy/parse_bam.h
@@ -7,7 +7,7 @@
 
 
 void cmap(char *bam, samFile *fp_in, char *chrom, int start, int end,
-				int min_mapq, int min_bseq, double *numer,
-				long *denom, int *count_mat, sam_hdr_t *bamHdr);
+				int min_mapq, int min_bseq, long double *numer,
+				unsigned long long *denom, int *count_mat, sam_hdr_t *bamHdr);
 
 int get_ref_length(sam_hdr_t *bamHdr, char *chrom);
